+++ ta4j-core/src/main/java/org/ta4j/core/indicators/RecentSwingHighIndicator.java
@@ -24,134 +24,163 @@
 package org.ta4j.core.indicators;
 
 import org.ta4j.core.BarSeries;
-import static org.ta4j.core.num.NaN.NaN;
+import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.helpers.HighPriceIndicator;
 import org.ta4j.core.num.Num;
+
+import static org.ta4j.core.num.NaN.NaN;
 
 /**
  * Recent Swing High Indicator.
  */
 public class RecentSwingHighIndicator extends CachedIndicator<Num> {
 
-    private final int surroundingLowerBars;
+    private final int precedingLowerBars;
+    private final int followingLowerBars;
     private final int allowedEqualBars;
+    Indicator<Num> indicator;
 
     /**
      * Constructs a RecentSwingHighIndicator
      *
-     * @param series               The BarSeries to be analyzed.
-     * @param surroundingLowerBars For a bar to be identified as a swing high, it
-     *                             must have a higher high than this number of bars
-     *                             both immediately preceding and immediately
-     *                             following
-     * @param allowedEqualBars     For a looser definition of peak, instead of
-     *                             requiring the surrounding bars to be strictly
-     *                             lower highs we allow this number of equal highs
-     *                             to either side (i.e. flat-ish peak or plateau)
-     * @throws IllegalArgumentException if surroundingLowerBars is less than or
-     *                                  equal to 0.
+     * @param indicator          The Indicator to be analyzed.
+     * @param precedingLowerBars For a bar to be identified as a swing high, it must
+     *                           have a higher high than this number of bars
+     *                           immediately preceding it.
+     * @param followingLowerBars For a bar to be identified as a swing high, it must
+     *                           have a higher high than this number of bars
+     *                           immediately following it.
+     * @param allowedEqualBars   For a looser definition of swing high, instead of
+     *                           requiring the surrounding bars to be strictly lower
+     *                           highs we allow this number of equal highs to either
+     *                           side (i.e. flat-ish valley)
+     * @throws IllegalArgumentException if precedingLowerBars is less than or equal
+     *                                  to 0.
+     * @throws IllegalArgumentException if followingLowerBars is less than 0.
      * @throws IllegalArgumentException if allowedEqualBars is less than 0.
      */
-    public RecentSwingHighIndicator(BarSeries series, int surroundingLowerBars, int allowedEqualBars) {
-        super(series);
+    public RecentSwingHighIndicator(Indicator<Num> indicator, int precedingLowerBars, int followingLowerBars,
+            int allowedEqualBars) {
+        super(indicator);
 
-        if (surroundingLowerBars <= 0) {
-            throw new IllegalArgumentException("surroundingLowerBars must be greater than 0");
+        if (precedingLowerBars <= 0) {
+            throw new IllegalArgumentException("precedingLowerBars must be greater than 0");
+        }
+        if (followingLowerBars < 0) {
+            throw new IllegalArgumentException("followingLowerBars must be 0 or greater");
         }
         if (allowedEqualBars < 0) {
             throw new IllegalArgumentException("allowedEqualBars must be 0 or greater");
         }
-        this.surroundingLowerBars = surroundingLowerBars;
+        this.precedingLowerBars = precedingLowerBars;
+        this.followingLowerBars = followingLowerBars;
         this.allowedEqualBars = allowedEqualBars;
+        this.indicator = indicator;
     }
 
     /**
-     * * Constructs a RecentSwingHighIndicator with the specified BarSeries and
-     * surrounding lower bars count and a default allowed equal bars count of 0
+     * Constructs a RecentSwingHighIndicator with the specified BarSeries (and
+     * defaulting to use HighPriceIndicator) and surrounding higher bars count and a
+     * default allowed equal bars of 0
      *
-     * @param series
-     * @param surroundingLowerBars
+     * @param series               The BarSeries to be analyzed.
+     * @param surroundingLowerBars The number of bars to consider on each side that
+     *                             must have lower highs to identify a swing high.
      */
     public RecentSwingHighIndicator(BarSeries series, int surroundingLowerBars) {
-        this(series, surroundingLowerBars, 0);
+        this(new HighPriceIndicator(series), surroundingLowerBars, surroundingLowerBars, 0);
     }
 
     /**
-     * Constructs a RecentSwingHighIndicator with the specified BarSeries, a default
-     * surrounding lower bars count of 2, and a default allowed equal bars count of
-     * 0
+     * Constructs a RecentSwingHighIndicator with the specified BarSeries (and
+     * defaulting to use HighPriceIndicator) and surrounding higher bars count
+     * defaulted to 3, and a default allowed equal bars of 0
      *
-     * @param series The BarSeries to be analyzed.
+     * @param series The BarSeries to be analyzed. must have lower highs to identify
+     *               a swing high.
      */
     public RecentSwingHighIndicator(BarSeries series) {
-        this(series, 2, 0);
+        this(new HighPriceIndicator(series), 3, 0, 0);
     }
 
-    /**
-     * Validates if the specified bar at currentIndex, considering the direction,
-     * meets the criteria for being a swing high.
-     *
-     * @param currentIndex The index of the current bar.
-     * @param direction    The direction for comparison (-1 for previous bars, 1 for
-     *                     following bars).
-     * @return true if the bar at currentIndex is a swing high considering the
-     *         specified direction; false otherwise.
-     */
-    private boolean validateBars(int currentIndex, int direction) {
-        Num currentHighPrice = getBarSeries().getBar(currentIndex).getHighPrice();
+    private boolean isSwingHigh(int index) {
+        Num currentPrice = this.indicator.getValue(index);
+
+        // Check bars before
+        if (!checkPrecedingBars(index, currentPrice)) {
+            return false;
+        }
+
+        // Check bars after (up to the current calculation index)
+        return checkFollowingBars(index, currentPrice);
+    }
+
+    private boolean checkPrecedingBars(int index, Num currentPrice) {
         int lowerBarsCount = 0;
         int equalBarsCount = 0;
 
-        for (int i = currentIndex + direction; i >= getBarSeries().getBeginIndex()
-                && i <= getBarSeries().getEndIndex(); i += direction) {
-            Num comparisonHighPrice = getBarSeries().getBar(i).getHighPrice();
+        for (int i = index - 1; i >= index - precedingLowerBars - allowedEqualBars && i >= 0; i--) {
+            Num comparisonPrice = this.indicator.getValue(i);
 
-            if (currentHighPrice.isEqual(comparisonHighPrice)) {
+            if (currentPrice.isEqual(comparisonPrice)) {
                 equalBarsCount++;
-
                 if (equalBarsCount > allowedEqualBars) {
                     return false;
                 }
-            } else if (currentHighPrice.isGreaterThan(comparisonHighPrice)) {
+            } else if (currentPrice.isLessThan(comparisonPrice)) {
+                return false;
+            } else {
                 lowerBarsCount++;
-                if (lowerBarsCount == surroundingLowerBars) {
+                if (lowerBarsCount == precedingLowerBars) {
                     return true;
                 }
-            } else {
-                break;
             }
         }
-        return false;
+
+        return lowerBarsCount == precedingLowerBars;
     }
 
-    /**
-     * Calculates the most recent swing high value up to the specified index.
-     *
-     * @param index The bar index up to which the calculation is done.
-     * @return The value of the most recent swing high if it exists; NaN otherwise.
-     */
+    private boolean checkFollowingBars(int index, Num currentPrice) {
+        int lowerBarsCount = 0;
+        int equalBarsCount = 0;
+
+        for (int i = index + 1; i < index + followingLowerBars + allowedEqualBars + 1
+                && i < getBarSeries().getBarCount(); i++) {
+            Num comparisonPrice = this.indicator.getValue(i);
+
+            if (currentPrice.isEqual(comparisonPrice)) {
+                equalBarsCount++;
+                if (equalBarsCount > allowedEqualBars) {
+                    return false;
+                }
+            } else if (currentPrice.isLessThan(comparisonPrice)) {
+                return false;
+            } else {
+                lowerBarsCount++;
+                if (lowerBarsCount == followingLowerBars) {
+                    return true;
+                }
+            }
+        }
+
+        return lowerBarsCount == followingLowerBars;
+    }
+
     @Override
     protected Num calculate(int index) {
-        if (index < surroundingLowerBars) {
+        if (index < getUnstableBars() || index >= getBarSeries().getBarCount()) {
             return NaN;
         }
 
         for (int i = index; i >= getBarSeries().getBeginIndex(); i--) {
-            if (validateBars(i, -1) && validateBars(i, 1)) {
-                return getBarSeries().getBar(i).getHighPrice();
+            if (isSwingHigh(i)) {
+                return this.indicator.getValue(i);
             }
         }
-
         return NaN;
     }
 
-    /**
-     * Returns the number of unstable bars as defined by the surroundingBars
-     * parameter.
-     *
-     * @return The number of unstable bars.
-     */
-    @Override
     public int getUnstableBars() {
-        return surroundingLowerBars;
+        return precedingLowerBars + followingLowerBars;
     }
 }