+++ ta4j-core/src/test/java/org/ta4j/core/indicators/RecentSwingLowIndicatorTest.java
@@ -25,19 +25,25 @@ package org.ta4j.core.indicators;
 
 import java.util.ArrayList;
 import java.util.List;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
+import org.ta4j.core.indicators.helpers.LowPriceIndicator;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.NaN;
 import org.ta4j.core.num.Num;
 
 import java.util.function.Function;
+
 import static org.junit.Assert.assertEquals;
 
 import org.ta4j.core.Bar;
+
 import static org.ta4j.core.TestUtils.assertNumEquals;
+
 import org.ta4j.core.mocks.MockBar;
 
 public class RecentSwingLowIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
@@ -80,7 +86,7 @@ public class RecentSwingLowIndicatorTest extends AbstractIndicatorTest<Indicator
         int surroundingBars = 2;
         RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(series, surroundingBars);
 
-        assertEquals(surroundingBars, swingLowIndicator.getUnstableBars());
+        assertEquals(surroundingBars * 2, swingLowIndicator.getUnstableBars());
     }
 
     @Test
@@ -93,10 +99,10 @@ public class RecentSwingLowIndicatorTest extends AbstractIndicatorTest<Indicator
 
     @Test
     public void testCalculate_With2SurroundingBarsAnd2AllowedEqualBars_ReturnsValue() {
-        RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(series, 2, 2);
+        RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(new LowPriceIndicator(series), 2, 2, 2);
 
-        assertNumEquals(8, swingLowIndicator.getValue(2));
-        assertNumEquals(8, swingLowIndicator.getValue(3));
+        assertNumEquals(NaN.NaN, swingLowIndicator.getValue(2));
+        assertNumEquals(NaN.NaN, swingLowIndicator.getValue(3));
         assertNumEquals(8, swingLowIndicator.getValue(4));
         assertNumEquals(8, swingLowIndicator.getValue(5));
         assertNumEquals(8, swingLowIndicator.getValue(6));
@@ -117,10 +123,10 @@ public class RecentSwingLowIndicatorTest extends AbstractIndicatorTest<Indicator
 
     @Test
     public void testCalculate_With2SurroundingBarsAnd1AllowedEqualBars_ReturnsValue() {
-        RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(series, 2, 1);
+        RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(new LowPriceIndicator(series), 2, 2, 3);
 
         assertNumEquals(NaN.NaN, swingLowIndicator.getValue(2));
-        assertNumEquals(8, swingLowIndicator.getValue(3));
+        assertNumEquals(NaN.NaN, swingLowIndicator.getValue(3));
         assertNumEquals(8, swingLowIndicator.getValue(4));
         assertNumEquals(8, swingLowIndicator.getValue(5));
         assertNumEquals(8, swingLowIndicator.getValue(6));
@@ -129,14 +135,14 @@ public class RecentSwingLowIndicatorTest extends AbstractIndicatorTest<Indicator
         assertNumEquals(6, swingLowIndicator.getValue(9));
         assertNumEquals(6, swingLowIndicator.getValue(10));
         assertNumEquals(6, swingLowIndicator.getValue(11));
-        assertNumEquals(6, swingLowIndicator.getValue(12));
-        assertNumEquals(6, swingLowIndicator.getValue(13));
-        assertNumEquals(6, swingLowIndicator.getValue(14));
-        assertNumEquals(6, swingLowIndicator.getValue(15));
-        assertNumEquals(6, swingLowIndicator.getValue(16));
-        assertNumEquals(6, swingLowIndicator.getValue(17));
-        assertNumEquals(6, swingLowIndicator.getValue(18));
-        assertNumEquals(6, swingLowIndicator.getValue(19));
+        assertNumEquals(5, swingLowIndicator.getValue(12));
+        assertNumEquals(5, swingLowIndicator.getValue(13));
+        assertNumEquals(5, swingLowIndicator.getValue(14));
+        assertNumEquals(5, swingLowIndicator.getValue(15));
+        assertNumEquals(5, swingLowIndicator.getValue(16));
+        assertNumEquals(5, swingLowIndicator.getValue(17));
+        assertNumEquals(5, swingLowIndicator.getValue(18));
+        assertNumEquals(5, swingLowIndicator.getValue(19));
     }
 
     @Test
@@ -187,6 +193,27 @@ public class RecentSwingLowIndicatorTest extends AbstractIndicatorTest<Indicator
         assertNumEquals(5, swingLowIndicator.getValue(19));
     }
 
+    @Test
+    public void testCalculate_OnMovingBarSeries_ReturnsValue() {
+        BarSeries movingSeries = new MockBarSeries(numFunction, 10); // movingSeries: [10]
+        ClosePriceIndicator closePrice = new ClosePriceIndicator(movingSeries);
+        RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(closePrice, 1, 1, 0);
+
+        movingSeries.addBar(new MockBar(movingSeries.getLastBar().getEndTime().plusDays(1), 8, numFunction)); // movingSeries:
+                                                                                                              // [10, 8]
+        assertNumEquals(NaN.NaN, swingLowIndicator.getValue(movingSeries.getEndIndex()));
+
+        movingSeries.addBar(new MockBar(movingSeries.getLastBar().getEndTime().plusDays(1), 5, numFunction)); // movingSeries:
+                                                                                                              // [10, 8,
+                                                                                                              // 5]
+        assertNumEquals(NaN.NaN, swingLowIndicator.getValue(movingSeries.getEndIndex()));
+
+        movingSeries.addBar(new MockBar(movingSeries.getLastBar().getEndTime().plusDays(1), 7, numFunction)); // movingSeries:
+                                                                                                              // [10, 8,
+                                                                                                              // 5, 7]
+        assertNumEquals(5, swingLowIndicator.getValue(movingSeries.getEndIndex()));
+    }
+
     @Test(expected = IllegalArgumentException.class)
     public void testConstructor_SurroundingBarsZero() {
         RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(series, 0);
@@ -194,7 +221,8 @@ public class RecentSwingLowIndicatorTest extends AbstractIndicatorTest<Indicator
 
     @Test(expected = IllegalArgumentException.class)
     public void testConstructor_AllowedEqualBarsNegative() {
-        RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(series, 1, -1);
+        RecentSwingLowIndicator swingLowIndicator = new RecentSwingLowIndicator(new LowPriceIndicator(series), 1, 1,
+                -1);
     }
 
 }